refactor: notJoinedClubNames에서 not 제거
