+++ frontend/src/constants/joinedClubNames.ts
